pwm, metronome

+++ adc.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <stdint.h>
+
 #define ADC_PIN 26
 #define MOVING_AVG_SIZE 16
 
-int map(int x);
+uint8_t map(uint16_t x);
 
 void init_adc();
 void scan_adc();
